requirements.txt
Django==4.2.17
google-generativeai==0.8.3
numpy==2.0.2
opencv-python==4.10.0.84
PyAutoGUI==0.9.54
python-chess==1.999
scikit-image==0.24.0

